backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
chromedriver-autoinstaller
click==8.1.7
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-rq==3.0.0
djangorestframework==3.15.2
filelock==3.13.1
frozendict==2.4.6
fsspec==2024.2.0
gunicorn
h11==0.14.0
html5lib==1.1
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
multitasking==0.0.11
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
outcome==1.3.0.post0
pandas==2.2.3
peewee==3.17.8
pillow==10.2.0
platformdirs==4.3.6
psycopg2-binary
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rq==2.0.0
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.2
sympy==1.13.1
torch
torchaudio
torchvision
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
yfinance==0.2.51
